+++ src/geocompy/apps/setmeasurement/__main__.py
@@ -1,0 +1,6 @@
+from .cli import build_parser
+
+
+parser = build_parser()
+args = parser.parse_args()
+args.func(args)